Move instrumentation.py out of cofacts_ai to adk/ top level
Instrumentation is cross-cutting infra, not specific to the cofacts_ai
agent — keeping it as a sibling of main.py means a future second agent
package can reuse it without reaching into another agent's package.
