+++ adk/main.py
@@ -6,7 +6,7 @@ from fastapi import Request
 from google.adk.cli.fast_api import get_fast_api_app
 
 from cofacts_ai.auth_context import cofacts_token_var
-from cofacts_ai.instrumentation import setup_instrumentation
+from instrumentation import setup_instrumentation
 
 _agents_dir = os.path.dirname(os.path.abspath(__file__))
 load_dotenv(os.path.join(_agents_dir, "cofacts_ai", ".env"))
